backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2